requirements.txt
python-dotenv
google-generativeai
google-ai-generativelanguage
google-auth
google-api-core
google-api-python-client

langchain
langchain-core
langchain_huggingface
langchain_chroma
langchain-google-genai
langchain-text-splitters

sentence-transformers
transformers
torch
numpy

chromadb
cryptography

pydantic
pydantic-settings
pydantic_core

requests
httpx
typing_extensions
tqdm

tokenizers
tiktoken
regex
rich
attrs
anyio
PyYAML
python-dateutil
pytz

streamlit